Varius Updates
+Added SM_Railing
+Added SM_StairRailing
+Added WallDoor Prefab
+Added Mannequin Prefab

Set Mannequin To Kinematic Body to prevent it being moved by the player.

Level1 Bottom floor now looks prettier.
Level1 Exit door leads to Level2.
Level2 Now added to build list.

CameraScript new has an offset.
Camera Offset set to 0, 1.5, 0.
Camera Offset allows the camera to be positioned above the player move making less of the focus on the floor.

Camera FOV changed to 90. This means that the camera should be positioned closer to the player (I found -6 To be a good range).