Started Level 12
+Added Projectile Script
+Added FireProjectile Script
+Added DartTrap Script
+Added Arrow Prefab

Started designing the ArrowTrap object. Currently fires a projectile with an inputted speed with an inputted delay.

The arrow has a initial lifespan that can be overwritten in the ArrowTrap.

Projectiles will always move forward until they hit an entity then they will be destroyed.

If the projectile doesn't hit anything for it's lifespan it will be destroyed.

*Tweeked Abilities to be more safe.

Models for these objects will come next update.